Refactor:  성분 분석 이미지 업로드 로직 커스텀훅으로 분리
